Add structures to store references to pages during memory deduplication.
At the beginning we only focus on 64-bit architectures which have a direct map
to physical pages, especially amd64. There are two structures needed to
reference pages - a structure for candidate pages to deduplicate and a structure
for already compared pages including deduplicated ones. For this purpose we use
a tail queue and a red-black tree.